balancer/app/requirements.txt
fastapi
pulp
pandas
requests
pydantic
uvicorn
unidecode
beautifulsoup4
tqdm
python-dotenv
psycopg2
openai